work_queue: start/stop controls for the worker
Since 'work_queue' is not the final class by design and it
has virtual functions creation of worker should be triggered
only after creation of the derived class. Otherwise data
race on vtable pointer possible and it is clearly detectable
with '-fsanitize=thread -fsanitize=undefined'.

To overcome this issue, special protected ctor in which
worker is not created automatically was introduced. Also
'start_worker'/'kill_worker' methods are introduced/revised.
They are both protected now.