Revert "Merge branch 'cmake-improvements'"
This reverts commit ad808062005313d421d11c650d049be7d8cf32a1, reversing
changes made to 498c6a1cee67024a7fa58bbb5d13cc14ea7659fa.

Custom 'check' target is not working how it is supposed to
in sub projects - it creates a mess with dependencies and
especially with the test logs. So for now this idea is not
acceptable and original CMake behavior when tests are build
together with the code is more preferable.